requirements.txt
flask
groq
reportlab
python-dotenv
requests